% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Database.R
\name{Database}
\alias{Database}
\title{Database}
\description{
Represents a database, handling the connections itself. It allows you to perform
operations through the use of functions (e.g. list tables, query data).
}
\section{Usage}{

\preformatted{
credentials <- DatabaseCredentials$new("connection", "user", "password")
database <- Database$new(credentials)

database$list(tableType)
database$listTableInformation(tableName)
database$query(query)
}
}

\section{Arguments}{

\describe{
  \item{credentials}{Object containing all the information required to connect to a database.}
  \item{dataSourceName}{Data structure normally used to describe the connection to a database.}
  \item{userID}{Id of the user, usually necessary to perform database connections.}
  \item{password}{Password of the user, usually necessary to perform database connections.}
  \item{tableType}{Table type (e.g. TABLE, VIEW)}
  \item{tableName}{Name of the table inside a database.}
  \item{query}{SQL query to run against the database.}
}
}

\section{Details}{

\code{$new(databaseCredentials = NULL)} Creates an instance of Database.

\code{$list(tableType = "TABLE")} List all the tables (name and description) in the database according to the type given.

\code{$listTableInformation(tableName = "")} List the information about a given table (names of the columns, their type and nullability).

\code{$query(query = NuLL)} Queries the database when given a valid SQL Query (e.g. SELECT * FROM table)
}

\examples{
\dontrun{
# Building the Database object.
dsn <- "connection"
uid <- "root"
pwd <- "password"

databaseCredentials <- DatabaseCredentials$new(dsn, uid, pwd)
database <- Database$new(databaseCredentials)

# Listing all the tables and views in the database.
database$list("TABLE")
database$list("VIEW")

# List information about a sepecif table.
database$listTableInformation("tableName")

# Query the database to get information.
database$query("SELECT * FROM table")
}

}
