% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatabaseCredentials.R
\name{DatabaseCredentials}
\alias{DatabaseCredentials}
\title{DatabaseCredentials}
\description{
Represents the credentials of a database which are necessary to perform a connection.
}
\section{Usage}{

\preformatted{
databaseCredentials <- DBCredentials$new(dataSourceName = "", userID = "", password = "")

databaseCredentials$getDataServiceName()
databaseCredentials$getUserID()
databaseCredentials$getPassword()
}
}

\section{Arguments}{

\describe{
  \item{dataSourceName}{Data structure normally used to describe the connection to a database.}
  \item{userID}{Id of the user, usually necessary to perform database connections.}
  \item{password}{Password of the user, usually necessary to perform database connections.}
}
}

\section{Details}{

\code{$new(dataSourceName = "", userID = "", password = "")} Creates an instance of DatabaseCredentials.

\code{$getDataSourceName()} Returns the data source name.

\code{$getUserID()} Returns the id of the user.

\code{$getPassword()} Returns the password of the user.
}

\examples{
# Building a simple DatabaseCredentials object and extracting each of its attributes.
dsn <- "connection"
uid <- "root"
pwd <- "password"

databaseCredentials <- DatabaseCredentials$new(dsn, uid, pwd)

databaseCredentials$getDataSourceName()
databaseCredentials$getUserID()
databaseCredentials$getPassword()

}
