% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQLQuery.R
\name{SQLQuery}
\alias{SQLQuery}
\title{SQLQuery}
\description{
This class represents an SQLQuery, storing inside it all the information regarding a
normal query. It can later be used to run against the database.
}
\section{Usage}{

\preformatted{
sql <- SQLQuery$new(query = "")

sql$toString()
}
}

\section{Arguments}{

\describe{
  \item{query}{Character representation of an SQL query.}
}
}

\section{Details}{

\code{$new(query = "")} Creates an instance of SQLQuery.

\code{$toString()} Returns the SQL query inside the object as a string.
}

\examples{
# Building a simple SQLQuery object and printing out its content.
query <- "SELECT * FROM tableName"
sqlQuery <- SQLQuery$new(query)

sqlQuery$toString()

}
